Module headless_lms_utils::error::backtrace_formatter

Custom formatting for stack backtraces inteded to be printed to the console for developers.

Functions

format_backtrace

Formats backtraces for printing but omits unnecessary stack frames.

print_filtered_frame_placeholder 🔒