Trait headless_lms_utils::file_store::FileStore

pub trait FileStore {
    // Required methods
    fn upload<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        path: &'life1 Path,
        contents: Vec<u8>,
        mime_type: &'life2 str
    ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn upload_stream<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        path: &'life1 Path,
        contents: GenericPayload,
        mime_type: &'life2 str
    ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn download<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = UtilResult<Vec<u8>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn download_stream<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = UtilResult<Box<dyn Stream<Item = Result<Bytes>>>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_direct_download_url<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = UtilResult<String>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_cache_files_folder_path(&self) -> UtilResult<&Path>;

    // Provided methods
    fn get_download_url(
        &self,
        path: &Path,
        app_conf: &ApplicationConfiguration
    ) -> String { ... }
    fn fetch_file_content_or_use_filesystem_cache<'life0, 'life1, 'async_trait>(
        &'life0 self,
        file_path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = UtilResult<Vec<u8>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Allows storing files to a file storage backend.

Required Methods

fn upload<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, path: &'life1 Path, contents: Vec<u8>, mime_type: &'life2 str ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Upload a file that’s in memory to a path.

fn upload_stream<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, path: &'life1 Path, contents: GenericPayload, mime_type: &'life2 str ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Upload a file without loading the whole file to memory

fn download<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 Path ) -> Pin<Box<dyn Future<Output = UtilResult<Vec<u8>>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Download a file to memory.

fn download_stream<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 Path ) -> Pin<Box<dyn Future<Output = UtilResult<Box<dyn Stream<Item = Result<Bytes>>>>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Download a file without loading the whole file to memory.

fn get_direct_download_url<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 Path ) -> Pin<Box<dyn Future<Output = UtilResult<String>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get a url that can be used to download the file without authentication for a while. In most cases you probably want to use get_download_url() instead.

fn delete<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 Path ) -> Pin<Box<dyn Future<Output = UtilResult<()>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Delete a file.

fn get_cache_files_folder_path(&self) -> UtilResult<&Path>

This function returns a path to a folder where downloaded files can be cached.

Provided Methods

fn get_download_url( &self, path: &Path, app_conf: &ApplicationConfiguration ) -> String

Get a url for a file in FileStore that can be used to access the resource.

fn fetch_file_content_or_use_filesystem_cache<'life0, 'life1, 'async_trait>( &'life0 self, file_path: &'life1 Path ) -> Pin<Box<dyn Future<Output = UtilResult<Vec<u8>>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,