headless_lms_utils

Module prelude

Re-exports

pub use crate::error::util_error::UtilError;

pub use crate::error::util_error::UtilErrorType;

pub use crate::error::util_error::UtilResult;

Modules

anyhow

github crates-io docs-rs

chrono

Chrono: Date and Time for Rust

futures

Abstractions for asynchronous programming.

serde_json

Serde JSON

tokio

A runtime for writing reliable network applications without compromising speed.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

info

Constructs an event at the info level.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

ApplicationConfiguration

AzureBlobStorageConfiguration

AzureChatbotConfiguration

AzureConfiguration

AzureSearchConfiguration

DateTime

ISO 8601 combined date and time with time zone.

OAuthServerConfiguration

Url

A parsed URL record.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Uuid

A Universally Unique Identifier (UUID).

Traits

BackendError

The error types of this program all implement this trait for interoperability.

Context

Provides the context method for Result.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Derive Macros

Deserialize

Serialize